api/requirements.txt

fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.17
python-dotenv==1.0.1
pymongo==4.10.1
bcrypt==4.2.0
PyJWT[crypto]==2.6.0
websockets